js confirmation when try to remove message

+++ hostme-web/src/main/webapp/WEB-INF/resources/js/conversations/conversation.js
@@ -106,7 +106,11 @@ function createChatItem(postDto) {
 function removePost() {
 	
 	var postId = $(this).attr('class').replace('deletePost ', '');
-	deleteMessage(postId);	
+	if (confirm('Do you really want to delete?')) {
+		deleteMessage(postId);
+    } else {
+        return false;
+    }
 }
 
 function sendMessage(conversationId, message) {
